+++ app/routers/health.ts
@@ -3,6 +3,7 @@
  *
  * @ignore
  */
+
 import * as express from 'express';
 const healthRouter = express.Router();
 
@@ -10,30 +11,74 @@ import * as createDebug from 'debug';
 import { OK } from 'http-status';
 import * as mongoose from 'mongoose';
 
-import mongooseConnectionOptions from '../../mongooseConnectionOptions';
+import * as redis from '../../redis';
 
-const debug = createDebug('sskts-api:healthRouter:health');
-const MONGOOSE_CONNECTION_READY_STATE_CONNECTED = 1;
+const debug = createDebug('sskts-api:healthRouter');
+// 接続確認をあきらめる時間(ミリ秒)
+const TIMEOUT_GIVE_UP_CHECKING_IN_MILLISECONDS = 3000;
 
 healthRouter.get(
     '',
-    (_, res, next) => {
-        debug('mongoose.connection.readyState is', mongoose.connection.readyState);
+    async (_, res, next) => {
+        try {
+            await Promise.all([
+                new Promise((resolve, reject) => {
+                    let givenUpChecking = false;
 
-        // mongodb接続状態チェック
-        if (mongoose.connection.readyState !== MONGOOSE_CONNECTION_READY_STATE_CONNECTED) {
-            debug('connecting mongodb...');
-            mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions, (err) => {
-                if (err instanceof Error) {
-                    next(err);
+                    // mongodb接続状態チェック
+                    mongoose.connection.db.admin().ping((err, result) => {
+                        debug('mongodb ping:', err, result);
+                        // すでにあきらめていたら何もしない
+                        if (givenUpChecking) {
+                            return;
+                        }
 
-                    return;
-                }
+                        if (err instanceof Error) {
+                            reject(err);
+                        } else {
+                            resolve();
+                        }
+                    });
 
-                res.status(OK).send('healthy!');
-            });
-        } else {
+                    setTimeout(
+                        () => {
+                            givenUpChecking = true;
+                            reject(new Error('unable to check db connection'));
+                        },
+                        TIMEOUT_GIVE_UP_CHECKING_IN_MILLISECONDS
+                    );
+                }),
+                new Promise(async (resolve, reject) => {
+                    let givenUpChecking = false;
+
+                    // redisサーバー接続が生きているかどうか確認
+                    redis.getClient().ping([], (err, reply) => {
+                        debug('redis ping:', err, reply);
+                        // すでにあきらめていたら何もしない
+                        if (givenUpChecking) {
+                            return;
+                        }
+
+                        if (err instanceof Error) {
+                            reject(err);
+                        } else {
+                            resolve();
+                        }
+                    });
+
+                    setTimeout(
+                        () => {
+                            givenUpChecking = true;
+                            reject(new Error('unable to check db connection'));
+                        },
+                        TIMEOUT_GIVE_UP_CHECKING_IN_MILLISECONDS
+                    );
+                })
+            ]);
+
             res.status(OK).send('healthy!');
+        } catch (error) {
+            next(error);
         }
     });