+++ loadtest/transactionByExistingReserve.js
@@ -1,4 +1,11 @@
 "use strict";
+/**
+ * すでに存在している座席予約を使った取引フローテストスクリプト
+ * 運用ではありえない状況
+ * シナリオテストに使うためのものです
+ *
+ * @ignore
+ */
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
@@ -8,84 +15,42 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-/**
- * すでに存在している座席予約を使った取引フローテストスクリプト
- * 運用ではありえない状況
- * シナリオテストに使うためのものです
- *
- * @ignore
- */
 const GMO = require("@motionpicture/gmo-service");
-const createDebug = require("debug");
 const httpStatus = require("http-status");
 const moment = require("moment");
 const request = require("request-promise-native");
 const winston = require("winston");
-const debug = createDebug('sskts-api:examples:transactionByExistingReserve');
+const processOneTransaction_1 = require("./scenarios/processOneTransaction");
+const repeatGMOAuthUntilSuccess_1 = require("./scenarios/repeatGMOAuthUntilSuccess");
+const wait_1 = require("./scenarios/wait");
 const logger = new (winston.Logger)({
     transports: [
         new (winston.transports.Console)({
-            timestamp: true,
-            level: 'info'
+            timestamp: false,
+            level: 'debug',
+            json: false,
+            showLevel: false
         }),
         new (winston.transports.File)({
-            filename: `transactionByExistingReserve-${moment().format('YYYYMMDDHHmmss')}.log`,
-            timestamp: true,
+            filename: `logs/transactionByExistingReserve-${moment().format('YYYYMMDDHHmmss')}.log`,
+            timestamp: false,
             level: 'info',
-            json: false
+            json: false,
+            showLevel: false
         })
     ]
 });
 const API_ENDPOINT = process.env.TEST_API_ENDPOINT;
-function tryAuthGMO(gmoShopId, gmoShopPass, amount) {
-    return __awaiter(this, void 0, void 0, function* () {
-        let result = null;
-        while (result === null) {
-            try {
-                // GMOオーソリ取得
-                const orderId = Date.now().toString();
-                const entryTranResult = yield GMO.CreditService.entryTran({
-                    shopId: gmoShopId,
-                    shopPass: gmoShopPass,
-                    orderId: orderId,
-                    jobCd: GMO.Util.JOB_CD_AUTH,
-                    amount: amount
-                });
-                const execTranResult = yield GMO.CreditService.execTran({
-                    accessId: entryTranResult.accessId,
-                    accessPass: entryTranResult.accessPass,
-                    orderId: orderId,
-                    method: '1',
-                    cardNo: '4111111111111111',
-                    expire: '2012',
-                    securityCode: '123'
-                });
-                debug(execTranResult);
-                result = {
-                    orderId: orderId,
-                    accessId: entryTranResult.accessId,
-                    accessPass: entryTranResult.accessPass
-                };
-            }
-            catch (error) {
-                debug(error);
-            }
-        }
-        return result;
-    });
-}
+const TEST_THEATER_ID = '118';
+// const TEST_GMO_SHOP_ID = 'tshop00026096';
+// const TEST_GMO_SHOP_PASS = 'xbxmkaa6';
+const TEST_GMO_SHOP_ID = 'tshop00026715';
+const TEST_GMO_SHOP_PASS = 'ybmbptww';
 // tslint:disable-next-line:max-func-body-length
-function main() {
+function main(coaSeatAuthorization, makeInrquiryResult) {
     return __awaiter(this, void 0, void 0, function* () {
-        let response;
-        const gmoShopId = 'tshop00026096';
-        const gmoShopPass = 'xbxmkaa6';
-        const theaterCode = '118';
-        const reserveNum = 1381;
-        const tel = '09012345678';
-        const totalPrice = 2600;
         // アクセストークン取得
-        response = yield request.post({
+        const accessToken = yield request.post({
             url: `${API_ENDPOINT}/oauth/token`,
             body: {
                 assertion: process.env.SSKTS_API_REFRESH_TOKEN,
@@ -94,12 +59,16 @@ function main() {
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('oauth token result:', response.statusCode, response.body);
+            return response.body.access_token;
         });
-        debug('oauth token result:', response.statusCode, response.body);
-        const accessToken = response.body.access_token;
+        const reserveNum = makeInrquiryResult.attributes.inquiry_key.reserve_num;
+        const tel = '09012345678';
+        const totalPrice = coaSeatAuthorization.price;
         // 取引開始
-        debug('starting transaction...');
-        response = yield request.post({
+        logger.debug('starting transaction...');
+        const startTransactionResult = yield request.post({
             url: `${API_ENDPOINT}/transactions/startIfPossible`,
             auth: { bearer: accessToken },
             body: {
@@ -108,16 +77,18 @@ function main() {
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('transaction start result:', response.statusCode, response.body);
+            if (response.statusCode === httpStatus.NOT_FOUND) {
+                throw new Error('please try later');
+            }
+            if (response.statusCode !== httpStatus.OK) {
+                handleError(response.body);
+            }
+            return response.body.data;
         });
-        debug('transaction start result:', response.statusCode, response.body);
-        if (response.statusCode === httpStatus.NOT_FOUND) {
-            throw new Error('please try later');
-        }
-        if (response.statusCode !== httpStatus.OK) {
-            throw new Error(response.body.message);
-        }
-        const transactionId = response.body.data.id;
-        const owners = response.body.data.attributes.owners;
+        const transactionId = startTransactionResult.id;
+        const owners = startTransactionResult.attributes.owners;
         const promoterOwner = owners.find((owner) => {
             return (owner.group === 'PROMOTER');
         });
@@ -126,112 +97,63 @@ function main() {
             return (owner.group === 'ANONYMOUS');
         });
         const anonymousOwnerId = (anonymousOwner) ? anonymousOwner.id : null;
+        // tslint:disable-next-line:no-magic-numbers
+        yield wait_1.default(5000);
         // COAオーソリ追加
-        debug('adding authorizations coaSeatReservation...');
-        response = yield request.post({
+        logger.debug('adding authorizations coaSeatReservation...');
+        const coaSeatAuthorizationNow = Object.assign({}, coaSeatAuthorization, { owner_to: anonymousOwnerId });
+        yield request.post({
             url: `${API_ENDPOINT}/transactions/${transactionId}/authorizations/coaSeatReservation`,
             auth: { bearer: accessToken },
-            body: {
-                seats: [
-                    {
-                        mvtk_num: '',
-                        mvtk_kbn_denshiken: '00',
-                        mvtk_kbn_maeuriken: '00',
-                        mvtk_kbn_kensyu: '00',
-                        mvtk_sales_price: 0,
-                        kbn_eisyahousiki: '00',
-                        add_glasses: 0,
-                        mvtk_app_price: 0,
-                        sale_price: 1300,
-                        dis_price: 0,
-                        add_price: 0,
-                        std_price: 1300,
-                        ticket_name_kana: 'レイトショー',
-                        ticket_name: {
-                            en: 'Late Show Price',
-                            ja: 'レイト'
-                        },
-                        ticket_code: '171',
-                        seat_code: 'Ｂ－３',
-                        section: '   ',
-                        performance: '11820170410162500902130'
-                    },
-                    {
-                        mvtk_num: '',
-                        mvtk_kbn_denshiken: '00',
-                        mvtk_kbn_maeuriken: '00',
-                        mvtk_kbn_kensyu: '00',
-                        mvtk_sales_price: 0,
-                        kbn_eisyahousiki: '00',
-                        add_glasses: 0,
-                        mvtk_app_price: 0,
-                        sale_price: 1300,
-                        dis_price: 0,
-                        add_price: 0,
-                        std_price: 1300,
-                        ticket_name_kana: 'レイトショー',
-                        ticket_name: {
-                            en: 'Late Show Price',
-                            ja: 'レイト'
-                        },
-                        ticket_code: '171',
-                        seat_code: 'Ｂ－４',
-                        section: '   ',
-                        performance: '11820170410162500902130'
-                    }
-                ],
-                owner_to: anonymousOwnerId,
-                owner_from: '5868e16789cc75249cdbfa4b',
-                price: totalPrice,
-                coa_screen_code: '90',
-                coa_time_begin: '2130',
-                coa_title_branch_num: '0',
-                coa_title_code: '16250',
-                coa_date_jouei: '20170410',
-                coa_theater_code: theaterCode,
-                coa_tmp_reserve_num: reserveNum
-            },
+            body: coaSeatAuthorizationNow,
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('addCOASeatReservationAuthorization result:', response.statusCode, response.body);
+            if (response.statusCode !== httpStatus.OK) {
+                handleError(response.body);
+            }
         });
-        debug('addCOASeatReservationAuthorization result:', response.statusCode, response.body);
-        if (response.statusCode !== httpStatus.OK) {
-            throw new Error(response.body.message);
-        }
+        // tslint:disable-next-line:no-magic-numbers
+        yield wait_1.default(5000);
         // GMOオーソリ取得(できるまで続ける)
-        const tryAuthGMOResult = yield tryAuthGMO(gmoShopId, gmoShopPass, totalPrice);
-        const orderId = tryAuthGMOResult.orderId;
-        const accessId = tryAuthGMOResult.accessId;
-        const accessPass = tryAuthGMOResult.accessPass;
+        const gmoAuthResult = yield repeatGMOAuthUntilSuccess_1.default({
+            gmoShopId: TEST_GMO_SHOP_ID,
+            gmoShopPass: TEST_GMO_SHOP_PASS,
+            amount: totalPrice,
+            orderIdPrefix: `${moment().format('YYYYMMDD')}${TEST_THEATER_ID}${moment().format('HHmmssSS')}`
+        });
+        logger.debug('gmo auth countTry:', gmoAuthResult.countTry);
         // GMOオーソリ追加
-        debug('adding authorizations gmo...');
-        response = yield request.post({
+        logger.debug('adding authorizations gmo...');
+        yield request.post({
             url: `${API_ENDPOINT}/transactions/${transactionId}/authorizations/gmo`,
             auth: { bearer: accessToken },
             body: {
                 owner_from: anonymousOwnerId,
                 owner_to: promoterOwnerId,
-                gmo_shop_id: gmoShopId,
-                gmo_shop_pass: gmoShopPass,
-                gmo_order_id: orderId,
+                gmo_shop_id: TEST_GMO_SHOP_ID,
+                gmo_shop_pass: TEST_GMO_SHOP_PASS,
+                gmo_order_id: gmoAuthResult.orderId,
                 gmo_amount: totalPrice,
-                gmo_access_id: accessId,
-                gmo_access_pass: accessPass,
+                gmo_access_id: gmoAuthResult.accessId,
+                gmo_access_pass: gmoAuthResult.accessPass,
                 gmo_job_cd: GMO.Util.JOB_CD_AUTH,
                 gmo_pay_type: GMO.Util.PAY_TYPE_CREDIT
             },
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('addGMOAuthorization result:', response.statusCode, response.body);
+            if (response.statusCode !== httpStatus.OK) {
+                handleError(response.body);
+            }
         });
-        debug('addGMOAuthorization result:', response.statusCode, response.body);
-        if (response.statusCode !== httpStatus.OK) {
-            throw new Error(response.body.message);
-        }
         // 購入者情報登録
-        debug('updating anonymous...');
-        response = yield request.patch({
+        logger.debug('updating anonymous...');
+        yield request.patch({
             url: `${API_ENDPOINT}/transactions/${transactionId}/anonymousOwner`,
             auth: { bearer: accessToken },
             body: {
@@ -242,32 +164,34 @@ function main() {
             },
             json: true,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('anonymousOwner updated.', response.statusCode, response.body);
+            if (response.statusCode !== httpStatus.NO_CONTENT) {
+                handleError(response.body);
+            }
         });
-        debug('anonymousOwner updated.', response.statusCode, response.body);
-        if (response.statusCode !== httpStatus.NO_CONTENT) {
-            throw new Error(response.body.message);
-        }
         // 照会情報登録(購入番号と電話番号で照会する場合)
-        debug('enabling inquiry...');
-        response = yield request.patch({
+        logger.debug('enabling inquiry...');
+        yield request.patch({
             url: `${API_ENDPOINT}/transactions/${transactionId}/enableInquiry`,
             auth: { bearer: accessToken },
             body: {
-                inquiry_theater: theaterCode,
+                inquiry_theater: TEST_THEATER_ID,
                 inquiry_id: reserveNum,
                 inquiry_pass: '00000000000' // 購入取消されないようにあえて間違った値
             },
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('enableInquiry result:', response.statusCode, response.body);
+            if (response.statusCode !== httpStatus.NO_CONTENT) {
+                handleError(response.body);
+            }
         });
-        debug('enableInquiry result:', response.statusCode, response.body);
-        if (response.statusCode !== httpStatus.NO_CONTENT) {
-            throw new Error(response.body.message);
-        }
         // メール追加
         const content = `
-sskts-api:examples:transactionByExistingReserve 様\n
+sskts-api:loadtest:transactionByExistingReserve 様\n
 \n
 -------------------------------------------------------------------\n
 この度はご購入いただき誠にありがとうございます。\n
@@ -285,8 +209,8 @@ sskts-api:examples:transactionByExistingReserve 様\n
 http://www.cinemasunshine.co.jp/\n
 -------------------------------------------------------------------\n
 `;
-        debug('adding email...');
-        response = yield request.post({
+        logger.debug('adding email...');
+        yield request.post({
             url: `${API_ENDPOINT}/transactions/${transactionId}/notifications/email`,
             auth: { bearer: accessToken },
             body: {
@@ -298,52 +222,113 @@ http://www.cinemasunshine.co.jp/\n
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('addEmail result:', response.statusCode, response.body);
+            if (response.statusCode !== httpStatus.OK) {
+                handleError(response.body);
+            }
         });
-        debug('addEmail result:', response.statusCode, response.body);
-        if (response.statusCode !== httpStatus.OK) {
-            throw new Error(response.body.message);
-        }
+        // tslint:disable-next-line:no-magic-numbers
+        yield wait_1.default(5000);
         // 取引成立
-        debug('closing transaction...');
-        response = yield request.patch({
+        logger.debug('closing transaction...');
+        yield request.patch({
             url: `${API_ENDPOINT}/transactions/${transactionId}/close`,
             auth: { bearer: accessToken },
             body: {},
             json: true,
             simple: false,
             resolveWithFullResponse: true
+        }).then((response) => {
+            logger.debug('close result:', response.statusCode, response.body);
+            if (response.statusCode !== httpStatus.NO_CONTENT) {
+                handleError(response.body);
+            }
         });
-        debug('close result:', response.statusCode, response.body);
-        if (response.statusCode !== httpStatus.NO_CONTENT) {
-            throw new Error(response.body.message);
-        }
+        return {
+            gmoAuthResult
+        };
     });
+}
+/**
+ * エラーレスポンスハンドリング
+ *
+ * @param {any} body レスポンスボディ
+ */
+function handleError(body) {
+    if (body.errors !== undefined) {
+        throw new Error(`${body.errors[0].title} ${body.errors[0].detail}`);
+    }
+    else {
+        throw new Error(body.text);
+    }
 }
 let count = 0;
 let numberOfClosedTransactions = 0;
 let numberOfProcessedTransactions = 0;
 const MAX_NUBMER_OF_PARALLEL_TASKS = 1800;
 const INTERVAL_MILLISECONDS = 500;
-const timer = setInterval(() => __awaiter(this, void 0, void 0, function* () {
-    if (count >= MAX_NUBMER_OF_PARALLEL_TASKS) {
-        clearTimeout(timer);
-        return;
-    }
-    count += 1;
-    const countNow = count;
-    try {
-        logger.info('starting...', countNow);
-        yield main();
-        numberOfClosedTransactions += 1;
-        logger.info('end', countNow);
-    }
-    catch (error) {
-        logger.error(error.message, countNow);
-    }
-    numberOfProcessedTransactions += 1;
-    logger.info('numberOfProcessedTransactions:', numberOfProcessedTransactions);
-    // count -= 1;
-}), INTERVAL_MILLISECONDS);
+const results = [];
+// まず普通にひとつの取引プロセス
+processOneTransaction_1.default({
+    apiEndpoint: API_ENDPOINT,
+    theaterId: TEST_THEATER_ID,
+    gmoShopId: TEST_GMO_SHOP_ID,
+    gmoShopPass: TEST_GMO_SHOP_PASS
+}).then((processTransactionResult) => {
+    logger.info('processTransaction success!', processTransactionResult);
+    const fields = ['no', 'started', 'processed', 'gmoOrderId', 'countTryGMOAuth', 'error'];
+    logger.info(`|${fields.join('|')}|\n|${Array.from(Array(fields.length)).map(() => ' :--- ').join('|')}|`);
+    const timer = setInterval(() => __awaiter(this, void 0, void 0, function* () {
+        if (count >= MAX_NUBMER_OF_PARALLEL_TASKS) {
+            clearTimeout(timer);
+            return;
+        }
+        count += 1;
+        const countNow = count;
+        const resultTransaction = {
+            no: countNow,
+            startedAt: new Date(),
+            processedAt: null,
+            error: '',
+            gmoOrderId: '',
+            countTryGMOAuth: 0
+        };
+        try {
+            logger.debug('starting...', countNow);
+            const mainResult = yield main(processTransactionResult.coaSeatAuthorization, processTransactionResult.makeInrquiryResult);
+            numberOfClosedTransactions += 1;
+            resultTransaction.countTryGMOAuth = mainResult.gmoAuthResult.countTry;
+            resultTransaction.gmoOrderId = mainResult.gmoAuthResult.orderId;
+        }
+        catch (error) {
+            resultTransaction.error = error.message;
+        }
+        numberOfProcessedTransactions += 1;
+        // 結果リストに追加
+        resultTransaction.processedAt = new Date();
+        results.push(resultTransaction);
+        // 結果リストをログ
+        logger.info(result2markdown(resultTransaction));
+    }), INTERVAL_MILLISECONDS);
+}).catch((err) => {
+    console.error(err);
+    process.exit(1);
+});
 process.on('exit', () => {
     logger.info('numberOfClosedTransactions:', numberOfClosedTransactions);
+    logger.info('results', [results]);
 });
+function result2markdown(resultTransaction) {
+    let markdown = '';
+    const resultStr = [
+        resultTransaction.no.toString(),
+        moment(resultTransaction.startedAt).format('YYYY-MM-DDTHH:mm:ss'),
+        (resultTransaction.processedAt instanceof Date) ? moment(resultTransaction.processedAt).format('YYYY-MM-DDTHH:mm:ss') : '',
+        resultTransaction.gmoOrderId,
+        resultTransaction.countTryGMOAuth,
+        resultTransaction.error
+    ].join('|');
+    markdown += `|${resultStr}|`;
+    return markdown;
+}